pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-assistant"
version = "0.1.0"
description = "Autonomous AI Research Assistant"
authors = [{name = "Research Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.29",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.0",
    "openai",
    "chromadb",
    "pymupdf",
    "tqdm",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "fitz>=0.0.1.dev2",
    "pathlib>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["packages*", "apps*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.json", "*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 